lugx-microservices/order-service/requirements.txt
flask
prometheus_flask_exporter
supabase==2.15.3
python-dotenv
prometheus_client
